requirements.txt
annotated-types==0.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
ffmpeg-python==0.2.0
filelock==3.15.4
fsspec==2024.6.1
future==1.0.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==24.1
preshed==3.0.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyspellchecker==0.8.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scipy==1.14.0
shellingham==1.5.4
smart-open==7.0.4
sounddevice==0.4.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.0
tbb==2021.13.0
thinc==8.2.5
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
wavio==0.0.9
weasel==0.4.1
wrapt==1.16.0
